+++ src/mutmut/workers/__init__.py
@@ -1,0 +1,1 @@
+"""Worker infrastructure for mutmut."""
